Refactor process data structure
