+++ include/process.h
@@ -6,15 +6,16 @@
 /*   By: aroque <aroque@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/06 17:45:59 by aroque            #+#    #+#             */
-/*   Updated: 2021/03/06 17:53:09 by aroque           ###   ########.fr       */
+/*   Updated: 2021/03/07 01:29:38 by aroque           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef COMMAND_H
-# define COMMAND_H
+#ifndef PROCESS_H
+# define PROCESS_H
 
 # include <stdbool.h>
 # include "libft.h"
+# include "token.h"
 
 typedef struct			s_file
 {
@@ -30,11 +31,12 @@ typedef struct			s_process
 	struct s_process	*next;
 }						t_process;
 
-int						parse_input_redirect(t_process *command, t_list **tk);
-int						parse_output_redirect(t_process *cmd,
-							t_list **tk, bool append);
+t_process				*parse_command(t_token **tokens);
+int						parse_input_redirect(t_process *process,
+							t_token **token);
+int						parse_output_redirect(t_process *process,
+							t_token **token, bool append);
 void					add_to_args(char **argv, char *word);
 void					push_process(t_process **lst, t_process *new);
-t_process				*parse_command(t_list **tokens);
 
 #endif
